requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
openai==1.54.0
numpy>=1.26.0
pypdf==5.1.0
python-docx==1.1.2
python-dotenv==1.0.1
tiktoken==0.12.0
